requirements.txt
waapi-client
pywin32
PyInstaller
platformdirs 